Address PR comments
Remove unused code.
Remove notion of Workspace next.
Change pull policy for plugin broker to Always.
