+++ assembly/assembly-wsmaster-war/src/main/java/org/eclipse/che/api/deploy/WsMasterModule.java
@@ -65,7 +65,7 @@ import org.eclipse.che.api.workspace.server.spi.provision.env.WorkspaceIdEnvVarP
 import org.eclipse.che.api.workspace.server.spi.provision.env.WorkspaceMavenServerJavaOptsEnvVariableProvider;
 import org.eclipse.che.api.workspace.server.stack.StackLoader;
 import org.eclipse.che.api.workspace.server.token.MachineTokenProvider;
-import org.eclipse.che.api.workspace.server.wsnext.WorkspaceNextApplier;
+import org.eclipse.che.api.workspace.server.wsplugins.ChePluginsApplier;
 import org.eclipse.che.commons.auth.token.ChainedTokenExtractor;
 import org.eclipse.che.commons.auth.token.RequestTokenExtractor;
 import org.eclipse.che.core.db.DBTermination;
@@ -266,7 +266,7 @@ public class WsMasterModule extends AbstractModule {
 
     bind(org.eclipse.che.api.user.server.AppStatesPreferenceCleaner.class);
 
-    MapBinder.newMapBinder(binder(), String.class, WorkspaceNextApplier.class);
+    MapBinder.newMapBinder(binder(), String.class, ChePluginsApplier.class);
   }
 
   private void configureSingleUserMode(Map<String, String> persistenceProperties) {
