+++ assembly/assembly-wsmaster-war/src/main/webapp/WEB-INF/classes/che/che.properties
@@ -485,12 +485,12 @@ che.singleport.wildcard_domain.ipless=false
 # Next properties are subject to changes and removal, so do not rely on them in a stable Che assembly
 
 # Workspace tooling plugins registry endpoint. Should be a valid HTTP URL that includes protocol, port etc.
-# In case Workspace.Next is not needed value 'NULL' should be used
+# In case Che plugins tooling is not needed value 'NULL' should be used
 che.workspace.feature.api=NULL
 
 # Docker image of Che plugin broker app that resolves workspace tooling configuration and copies
 # plugins dependencies to a workspace
-che.workspace.plugin_broker.image=garagatyi/broker:websocket
+che.workspace.plugin_broker.image=eclipse/che-plugin-broker:latest
 
 # Configures in which way secure servers will be protected with authentication.
 # Suitable values:
