+++ multiuser/integration-tests/che-multiuser-cascade-removal/src/test/java/org/eclipse/che/multiuser/integration/jpa/cascaderemoval/JpaEntitiesCascadeRemovalTest.java
@@ -85,7 +85,7 @@ import org.eclipse.che.api.workspace.server.spi.RuntimeInfrastructure;
 import org.eclipse.che.api.workspace.server.spi.StackDao;
 import org.eclipse.che.api.workspace.server.spi.WorkspaceDao;
 import org.eclipse.che.api.workspace.server.spi.environment.InternalEnvironmentFactory;
-import org.eclipse.che.api.workspace.server.wsnext.WorkspaceNextApplier;
+import org.eclipse.che.api.workspace.server.wsplugins.ChePluginsApplier;
 import org.eclipse.che.commons.env.EnvironmentContext;
 import org.eclipse.che.commons.lang.Pair;
 import org.eclipse.che.commons.subject.SubjectImpl;
@@ -269,7 +269,7 @@ public class JpaEntitiesCascadeRemovalTest {
                 bind(String.class)
                     .annotatedWith(Names.named("che.workspace.feature.api"))
                     .toInstance("");
-                MapBinder.newMapBinder(binder(), String.class, WorkspaceNextApplier.class);
+                MapBinder.newMapBinder(binder(), String.class, ChePluginsApplier.class);
                 Multibinder.newSetBinder(binder(), ResourceType.class)
                     .addBinding()
                     .to(RamResourceType.class);
