--- wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/wsnext/model/Command.java
@@ -1,110 +1,0 @@
-/*
- * Copyright (c) 2012-2018 Red Hat, Inc.
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- */
-package org.eclipse.che.api.workspace.server.wsnext.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-public class Command {
-
-  private String name = null;
-
-  @JsonProperty("working-dir")
-  private String workingDir = null;
-
-  private List<String> command = new ArrayList<String>();
-
-  /** */
-  public Command name(String name) {
-    this.name = name;
-    return this;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  /** */
-  public Command workingDir(String workingDir) {
-    this.workingDir = workingDir;
-    return this;
-  }
-
-  public String getWorkingDir() {
-    return workingDir;
-  }
-
-  public void setWorkingDir(String workingDir) {
-    this.workingDir = workingDir;
-  }
-
-  /** */
-  public Command command(List<String> command) {
-    this.command = command;
-    return this;
-  }
-
-  public List<String> getCommand() {
-    return command;
-  }
-
-  public void setCommand(List<String> command) {
-    this.command = command;
-  }
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Command command = (Command) o;
-    return Objects.equals(name, command.name)
-        && Objects.equals(workingDir, command.workingDir)
-        && Objects.equals(command, command.command);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(name, workingDir, command);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Command {\n");
-
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    workingDir: ").append(toIndentedString(workingDir)).append("\n");
-    sb.append("    command: ").append(toIndentedString(command)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
