--- wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/wsnext/model/Volume.java
@@ -1,86 +1,0 @@
-/*
- * Copyright (c) 2012-2018 Red Hat, Inc.
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- */
-package org.eclipse.che.api.workspace.server.wsnext.model;
-
-import java.util.Objects;
-
-public class Volume {
-
-  private String name = null;
-  private String mountPath = null;
-
-  /** */
-  public Volume name(String name) {
-    this.name = name;
-    return this;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  /** mountPath of the volume in running container */
-  public Volume mountPath(String path) {
-    this.mountPath = path;
-    return this;
-  }
-
-  public String getMountPath() {
-    return mountPath;
-  }
-
-  public void setMountPath(String mountPath) {
-    this.mountPath = mountPath;
-  }
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Volume volume = (Volume) o;
-    return Objects.equals(name, volume.name) && Objects.equals(mountPath, volume.mountPath);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(name, mountPath);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class Volume {\n");
-
-    sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    mountPath: ").append(toIndentedString(mountPath)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
